Move current IAM job into specific package
