+++ hq/app/AppComponents.scala
@@ -27,7 +27,8 @@ import play.api.routing.Router
 import play.api.{BuiltInComponentsFromContext, Logging}
 import play.filters.csrf.CSRFComponents
 import router.Routes
-import schedule.{Dynamo, IamJob, JobScheduler}
+import schedule.vulnerable.IamVulnerableUserJob
+import schedule.{Dynamo, JobScheduler}
 import services.{CacheService, MetricService}
 import utils.attempt.Attempt
 
@@ -148,7 +149,7 @@ class AppComponents(context: Context)
 
   //initialise IAM notification service
   val quartzScheduler = StdSchedulerFactory.getDefaultScheduler
-  val iamJob = new IamJob(enabled = true, cacheService, securitySnsClient, dynamo, configuration, iamClients)(executionContext)
+  val iamJob = new IamVulnerableUserJob(enabled = true, cacheService, securitySnsClient, dynamo, configuration, iamClients)(executionContext)
   val jobScheduler = new JobScheduler(quartzScheduler, List(iamJob))
   //jobScheduler.initialise() TODO: uncomment when ready to release creds reaper feature in PROD
