+++ hq/app/controllers/CredentialsController.scala
@@ -11,13 +11,14 @@ import play.api._
 import play.api.libs.json.{Json, OWrites, Writes}
 import play.api.libs.ws.WSClient
 import play.api.mvc._
-import schedule.{Dynamo, IamJob}
+import schedule.Dynamo
+import schedule.vulnerable.IamVulnerableUserJob
 import services.CacheService
 import utils.attempt.PlayIntegration.attempt
 
 import scala.concurrent.ExecutionContext
 
-class CredentialsController(val config: Configuration, cacheService: CacheService, val authConfig: GoogleAuthConfig, val iamJob: IamJob, val configuration: Configuration, val dynamo: Dynamo)
+class CredentialsController(val config: Configuration, cacheService: CacheService, val authConfig: GoogleAuthConfig, val iamJob: IamVulnerableUserJob, val configuration: Configuration, val dynamo: Dynamo)
                            (implicit val ec: ExecutionContext, val wsClient: WSClient, val bodyParser: BodyParser[AnyContent], val controllerComponents: ControllerComponents, val assetsFinder: AssetsFinder)
   extends BaseController  with SecurityHQAuthActions {
